requirements.txt
astroid==2.4.2
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.2.0
pylint==2.5.3
six==1.15.0
toml==0.10.1
wrapt==1.12.1
